+++ src/lib.rs
@@ -1,9 +1,10 @@
-//! A simplified TimeWarrior client written in Rust.
+//! Command-line interface (CLI) time tracker.
 //!
-//! This project is for learning purpose only.
-//! If you need to track your time consider using TimeWarrior.
+//! This project is for educational purpose only.
 //!
-//! Dates and Time are provided by the `chrono` crate
+//! It is a _partial_ Rust implementation of [Timewarrior](https://github.com/GothenburgBitFactory/timewarrior).
+//!
+//! For a stable feature-rich CLI time tracker, please use Timewarrior: <https://timewarrior.net/>.
 
 use anyhow::anyhow;
 use chrono::{DateTime, Local};
@@ -26,20 +27,20 @@ pub type Tags = Vec<Tag>;
 /// `ActivityId` = `usize`
 pub type ActivityId = usize;
 
-/// New Type on chrono::Date<Local>
+/// Newtype on `chrono::Date<Local>`
 ///
 /// Date is given in local time for convenience
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-pub struct AbsTime(DateTime<Local>);
+pub struct DateTimeW(DateTime<Local>);
 
-/// Convert from `DateTime<Local>` to `AbsTime`
-impl Into<AbsTime> for DateTime<Local> {
-    fn into(self) -> AbsTime {
-        AbsTime(self)
+/// Convert from `DateTime<Local>` to `DateTimeW`
+impl Into<DateTimeW> for DateTime<Local> {
+    fn into(self) -> DateTimeW {
+        DateTimeW(self)
     }
 }
 
-impl std::ops::Sub for AbsTime {
+impl std::ops::Sub for DateTimeW {
     type Output = DurationW;
 
     fn sub(self, rhs: Self) -> Self::Output {
@@ -47,13 +48,13 @@ impl std::ops::Sub for AbsTime {
     }
 }
 
-impl std::fmt::Display for AbsTime {
+impl std::fmt::Display for DateTimeW {
     fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), Error> {
         write!(f, "{}", self.0.format(DATETIME_FMT))
     }
 }
 
-/// New Type on `chrono::Duration`
+/// Newtype on `chrono::Duration`
 pub struct DurationW(chrono::Duration);
 
 impl fmt::Display for DurationW {
@@ -68,60 +69,62 @@ impl fmt::Display for DurationW {
     }
 }
 
-/// Time
+/// Time (absolute or relative)
+#[non_exhaustive]
 #[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
 pub enum Time {
-    /// Now, can be converted to `AbsTime` using `Clock.abs_time`
+    /// Now, can be converted to `DateTimeW` using `Clock.date_time`
     Now,
+    /// MinutesAgo, can be converted to `DateTimeW` using `Clock.date_time`
     MinutesAgo(usize),
-    Abs(AbsTime),
+    DateTime(DateTimeW),
 }
 
 /// Clock Abstraction
 pub trait Clock {
     /// Get current local time
-    fn get_time(&self) -> AbsTime;
-    /// Convert Time to absolute time
+    fn get_time(&self) -> DateTimeW;
+    /// Convert a `Time` to absolute time
     ///
-    /// `clock.abs_time(Time::Now)` equals approximately clock.get_time();
-    fn abs_time(&self, time: Time) -> AbsTime;
+    /// `clock.date_time(Time::Now)` equals approximately clock.get_time();
+    fn date_time(&self, time: Time) -> DateTimeW;
 
     /// Get time range for today
     ///
     /// today: 00:00:00 - 23:59:59
-    fn today_range(&self) -> (AbsTime, AbsTime);
+    fn today_range(&self) -> (DateTimeW, DateTimeW);
 
     /// Get time range for yesterday
     ///
     /// yesterday: 00:00:00 - 23:59:59
-    fn yesterday_range(&self) -> (AbsTime, AbsTime);
+    fn yesterday_range(&self) -> (DateTimeW, DateTimeW);
 
     /// Get time range for last week
     ///
     /// last week (ISO 8601, week start on monday)
     ///
     /// last week: monday: 00:00:00 - sunday: 23:59:59
-    fn last_week_range(&self) -> (AbsTime, AbsTime);
+    fn last_week_range(&self) -> (DateTimeW, DateTimeW);
 }
 
-/// A finished Activity (with a stop_time)
+/// A finished activity (with a stop time)
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct Activity {
     /// Activity start time
-    start_time: AbsTime,
+    start_time: DateTimeW,
     /// Activity `stop time` >= `start time`
-    stop_time: AbsTime,
+    stop_time: DateTimeW,
     /// Activity tags
     tags: Tags,
 }
 
 impl Activity {
     /// start time getter
-    pub fn get_start_time(&self) -> AbsTime {
+    pub fn get_start_time(&self) -> DateTimeW {
         self.start_time
     }
     /// stop time getter
-    pub fn get_stop_time(&self) -> AbsTime {
+    pub fn get_stop_time(&self) -> DateTimeW {
         self.stop_time
     }
     /// Return activity duration
@@ -138,6 +141,7 @@ impl Activity {
     }
 }
 
+/// Activities are sorted by start time
 impl Ord for Activity {
     fn cmp(&self, other: &Self) -> Ordering {
         self.get_start_time().cmp(&other.get_start_time())
@@ -150,29 +154,29 @@ impl PartialOrd for Activity {
     }
 }
 
-/// A started and unfinished Activity (no stop time)
+/// A started and unfinished activity (no stop time)
 #[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ActiveActivity {
+pub struct OngoingActivity {
     /// start time
-    pub start_time: AbsTime,
+    pub start_time: DateTimeW,
     /// Activity tags
     pub tags: Tags,
 }
 
-impl ActiveActivity {
+impl OngoingActivity {
     /// Constructor
-    pub fn new(start_time: AbsTime, tags: Tags) -> Self {
-        ActiveActivity { start_time, tags }
+    pub fn new(start_time: DateTimeW, tags: Tags) -> Self {
+        OngoingActivity { start_time, tags }
     }
     /// Start time getter
-    pub fn get_start_time(&self) -> AbsTime {
+    pub fn get_start_time(&self) -> DateTimeW {
         self.start_time
     }
     /// Return title (activity tags joined by a space)
     ///
     /// ```
-    /// # use rtw::{ActiveActivity, AbsTime};
-    /// let activity = ActiveActivity::new(chrono::Local::now().into(), vec![String::from("foo"), String::from("bar")]);
+    /// # use rtw::{OngoingActivity, DateTimeW};
+    /// let activity = OngoingActivity::new(chrono::Local::now().into(), vec![String::from("foo"), String::from("bar")]);
     /// assert_eq!(activity.get_title(), "foo bar");
     /// ```
     pub fn get_title(&self) -> String {
@@ -181,12 +185,12 @@ impl ActiveActivity {
     /// Convert active activity to finished activity
     ///
     /// ```
-    /// # use rtw::{ActiveActivity, AbsTime, Activity};
-    /// let activity = ActiveActivity::new(chrono::Local::now().into(), vec![String::from("foo"), String::from("bar")]);
+    /// # use rtw::{OngoingActivity, DateTimeW, Activity};
+    /// let activity = OngoingActivity::new(chrono::Local::now().into(), vec![String::from("foo"), String::from("bar")]);
     /// let finished: Activity = activity.into_activity(chrono::Local::now().into()).unwrap();
     /// ```
     /// `stop_time` should be >= `start_time` otherwise error
-    pub fn into_activity(self, stop_time: AbsTime) -> anyhow::Result<Activity> {
+    pub fn into_activity(self, stop_time: DateTimeW) -> anyhow::Result<Activity> {
         if self.start_time <= stop_time {
             Ok(Activity {
                 start_time: self.start_time,
@@ -203,26 +207,26 @@ impl ActiveActivity {
     }
 }
 
-/// A service for Activities
+/// A service for activities
 ///
 /// Abstracts activities queries and modifications
 pub trait ActivityService {
     /// Get current activity if any
     ///
     /// May fail depending on backend implementation
-    fn get_current_activity(&self) -> anyhow::Result<Option<ActiveActivity>>;
+    fn get_current_activity(&self) -> anyhow::Result<Option<OngoingActivity>>;
     /// Start a new activity
     ///
     /// May fail depending on backend implementation
     ///
     /// Returns new current activity
-    fn start_activity(&mut self, activity: ActiveActivity) -> anyhow::Result<ActiveActivity>;
+    fn start_activity(&mut self, activity: OngoingActivity) -> anyhow::Result<OngoingActivity>;
     /// Stop current activity
     ///
     /// May fail depending on backend implementation
     ///
     /// Returns stopped activity if any
-    fn stop_current_activity(&mut self, time: AbsTime) -> anyhow::Result<Option<Activity>>;
+    fn stop_current_activity(&mut self, time: DateTimeW) -> anyhow::Result<Option<Activity>>;
     /// Filter finished activities
     ///
     /// May fail depending on implementation
@@ -235,9 +239,9 @@ pub trait ActivityService {
         P: Fn(&(ActivityId, Activity)) -> bool;
 }
 
-/// A service for persisting finished activities
+/// A service for persisting and querying finished activities
 ///
-/// Abstracts finished activities persistence
+/// Abstracts finished activities queries and persistence
 pub trait FinishedActivityRepository {
     /// Write finished activity
     ///
@@ -255,18 +259,18 @@ pub trait FinishedActivityRepository {
         P: Fn(&(ActivityId, Activity)) -> bool;
 }
 
-/// A service for persisting current activity
+/// A service for persisting and querying current activity
 ///
-/// Abstracts current activity persistence
+/// Abstracts current activity queries and persistence
 pub trait CurrentActivityRepository {
     /// Retrieve current activity if any
     ///
     /// May fail depending on backend implementation
-    fn get_current_activity(&self) -> anyhow::Result<Option<ActiveActivity>>;
+    fn get_current_activity(&self) -> anyhow::Result<Option<OngoingActivity>>;
     /// Set `activity` as current activity
     ///
     /// May fail depending on backend implementation
-    fn set_current_activity(&mut self, activity: ActiveActivity) -> anyhow::Result<()>;
+    fn set_current_activity(&mut self, activity: OngoingActivity) -> anyhow::Result<()>;
     /// Reset current activity to none
     ///
     /// After calling this function, get_current_activity should return None