+++ CHANGELOG.md
@@ -18,6 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * github action
 * badges
 
+### Changed
+
+* This release requires [Rust 1.40](https://github.com/rust-lang/rust/blob/master/RELEASES.md#version-1400-2019-12-19)
+* `AbsTime` renamed to `DateTimeW`
+* `ActiveActivity` renamed to `OngoingActivity`
+
 ### Fixed
 
 * `summary` output is now sorted by start date